pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawsafe"
version = "0.1.0"
description = "The AI agent safety firewall — reversibility enforcement for OpenClaw and MCP agents"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "ClawSafe" }]
keywords = ["ai", "agent", "safety", "openclaw", "mcp", "firewall"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
cloud = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "anthropic>=0.40.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.9.0",
]
notifications = [
    "python-telegram-bot>=21.0.0",
    "resend>=0.8.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[project.scripts]
clawsafe = "clawsafe.cli:main"

[project.urls]
Homepage = "https://clawsafe.dev"
Repository = "https://github.com/clawsafe/clawsafe"

[tool.hatch.build.targets.wheel]
packages = ["clawsafe"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
